Give additional functionality to App component
I've made it so that, after adding an item to the cart, the new
maximum value of items that can be selected for any given item in the
cart is the maximum value minus the number in the cart. For example,
an item with a maximum 100 units available with 10 already in the cart
will only allow 90 to be selected for that item card.