Fix bug with ShopItemCard integer input
There was an issue where, if the quantity in the ShopItemCard was added
to the cart and then as a result that value exceeded the max remaining
availablility of that item, while the input would adjust accordingly and
show the new maximum, reducing the number of items in the cart for that
item would increase the quantity in the ShopItemCard's quantity input
until it reached the amount just added to the cart. This was due to it
taking the currentQuantity value from the items state value, which was
not being updated when the quantity of the item in the cart was changed.
Updating the items state fixed the issue.