Fix IntegerInput component
The IntegerInput component's value was being wholly controlled by the
onChange callback function passed into it, meaning I was unable to test
a change in the input's value using userEvent.type unless I was
deliberately passing in a callback function that did what I wanted it
to do. The problem with this is it doesn't account for every scenario
where that callback is fired; for example: some callbacks may not change
the input's value at all. Now, the IntegerInput component uses state to
control its own value while simultaneously passing that new value as the
only argument to the onChange callback function, meaning we can still
read the new value and trigger a manual re-render from a parent
component if necessary. A useEffect hook within IntegerInput then
listens for a change in the 'integer' prop's corrected value and can
trigger a re-render that way by then setting state within IntegerInput
itself when that value changes.